// amplify/backend/function/scriptGraphQL/src/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type CreateSheetInput = {
  companyID: string,
  year: number,
  grade: string,
  careerPlan?: string | null,
  careerPlanComment?: string | null,
  reviewComment?: string | null,
  reviewDate?: string | null,
  selfCheckDate?: string | null,
  firstComment?: string | null,
  firstCheckDate?: string | null,
  secondComment?: string | null,
  secondCheckDate?: string | null,
  overAllEvaluation?: number | null,
  statusValue?: number | null,
  interviewPlanDate?: string | null,
  interviewPlanComment?: string | null,
  InterviewMid1Date?: string | null,
  InterviewMid1Comment?: string | null,
  InterviewMid2Date?: string | null,
  InterviewMid2Comment?: string | null,
  InterviewMid3Date?: string | null,
  InterviewMid3Comment?: string | null,
  revieweeUsername: string,
  secondUsername: string,
  sheetGroupLocalId: string,
  sheetGroupName?: string | null,
  referencer?: Array< string > | null,
  reviewee: string,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
};

export type ModelSheetConditionInput = {
  grade?: ModelStringInput | null,
  careerPlan?: ModelStringInput | null,
  careerPlanComment?: ModelStringInput | null,
  reviewComment?: ModelStringInput | null,
  reviewDate?: ModelStringInput | null,
  selfCheckDate?: ModelStringInput | null,
  firstComment?: ModelStringInput | null,
  firstCheckDate?: ModelStringInput | null,
  secondComment?: ModelStringInput | null,
  secondCheckDate?: ModelStringInput | null,
  overAllEvaluation?: ModelIntInput | null,
  statusValue?: ModelIntInput | null,
  interviewPlanDate?: ModelStringInput | null,
  interviewPlanComment?: ModelStringInput | null,
  InterviewMid1Date?: ModelStringInput | null,
  InterviewMid1Comment?: ModelStringInput | null,
  InterviewMid2Date?: ModelStringInput | null,
  InterviewMid2Comment?: ModelStringInput | null,
  InterviewMid3Date?: ModelStringInput | null,
  InterviewMid3Comment?: ModelStringInput | null,
  revieweeUsername?: ModelIDInput | null,
  secondUsername?: ModelIDInput | null,
  sheetGroupLocalId?: ModelIDInput | null,
  sheetGroupName?: ModelStringInput | null,
  and?: Array< ModelSheetConditionInput | null > | null,
  or?: Array< ModelSheetConditionInput | null > | null,
  not?: ModelSheetConditionInput | null,
};
// ...
export type CreateSheetMutationVariables = {
  input?: CreateSheetInput,
  condition?: ModelSheetConditionInput | null,
};